requirements.txt
streamlit
google-generativeai
pinecone
torch
transformers
numpy
scikit-learn